% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_usq.R
\name{barplot_usq}
\alias{barplot_usq}
\title{Basic barplots using USQ colours and typography}
\usage{
barplot_usq(height, col = "usq charcoal", ...)
}
\arguments{
\item{height}{Either a vector or matrix of values describing the bars which
make up the plot. If height is a vector, the plot consists of a sequence of
rectangular bars with heights given by the values in the vector. If height is
a matrix and beside is \code{FALSE} then each bar of the plot corresponds to
a column of height, with the values in the column giving the heights of
stacked sub-bars making up the bar. If height is a matrix and beside is
\code{TRUE}, then the values in each column are juxtaposed rather than
stacked.}

\item{col}{Colour to use as fill for the bars as a digit. Defaults to 1, "usq
charcoal". There are six colours that can be used, see \strong{Details} for
more.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics]{par}}).}
}
\description{
Basic barplots that follow USQ colour and typography
guidelines that uses (hopefully) sensible defaults.
}
\details{
Names of colours for use in plots from the USQ Visual Identity Palette are
available from \code{link{usq_cols}}.
}
\examples{

barplot_usq(islands, col = "secondary orange")

}
