% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_usq.R
\name{boxplot_usq}
\alias{boxplot_usq}
\title{Basic Boxplots Using USQ Colours and Typography}
\usage{
boxplot_usq(x, main = "", sub = "", xlab = "", ylab = "",
  col = "usq charcoal", ...)
}
\arguments{
\item{x}{for specifying data from which the boxplots are to be produced.
Either a numeric vector, or a single list containing such vectors. Additional
unnamed arguments specify further data as separate vectors (each
corresponding to a component boxplot). \code{NA}s are allowed in the data.}

\item{main}{Main title. Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label. Optional, if not supplied it will be
blank.}

\item{xlab}{X-axis label. Optional.}

\item{ylab}{Y-axis label. Optional.}

\item{col}{Colour to use as fill and outline for boxes and outlier points as
a digit. Defaults to "usq charcoal". There are fifteen colours that can be
used, see \strong{Details} for more.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics]{par}}).}
}
\description{
Basic boxplots that follow USQ colour and typography
guidelines that uses (hopefully) sensible defaults.
}
\details{
Names of colours for use in plots from the USQ Visual Identity Palette are
available from \code{link{usq_cols}}.
}
\examples{
# plot using "secondary orange" as the fill colour

rb <- boxplot_usq(decrease ~ treatment,
                  data = OrchardSprays,
                  col = "secondary orange")
title("Comparing boxplot()s and non-robust mean +/- SD")

}
\seealso{
\code{\link[graphics]{boxplot}} for full documentation of the basic
boxplot capabilities
}
