% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_usq.R
\name{hist_usq}
\alias{hist_usq}
\title{Basic Histograms Using USQ Colours and Typography}
\usage{
hist_usq(x, main = "", sub = "", xlab = "", ylab = "Count",
  breaks = "pretty", col = "usq charcoal", ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{main}{Main title. Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label. Optional, if not supplied it will be
blank.}

\item{xlab}{X-axis label. Optional, if not supplied will be taken from data.}

\item{ylab}{Y-axis label. Optional, if not supplied will default to "Count".}

\item{breaks}{Either \code{pretty}, default, or \code{exact}. Pretty uses
Scott's Rule, whereas exact uses a bin-width of 1 so each value is
represented individually with a bar.}

\item{col}{Colour to use as fill and outline for boxes and outlier points as
a digit. Defaults to "usq charcoal". There are fifteen colours that can be
used, see \strong{Details} for more.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics]{par}}).}
}
\description{
Basic histograms that follow USQ colour and typography
guidelines that uses (hopefully) sensible defaults.
}
\details{
Names of colours for use in plots from the USQ Visual Identity Palette are
available from \code{link{usq_cols}}.
}
\examples{

hist_usq(islands)

}
\references{
Scott, D.W. (1979) On optimal and data-based histograms. \emph{Biometrika},
66, 605–610.

Hyndman, R. J., The Problem with Sturges’ Rule for Constructing Histograms,
1995, Monash University,
\url{https://www.robjhyndman.com/papers/sturges.pdf}.
}
\seealso{
\code{\link[graphics]{hist}} for full documentation of the basic
histogram capabilities
}
