% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_usq.R
\name{plot_usq}
\alias{plot_usq}
\title{Basic X-Y Plotting Using USQ Colours and Typography}
\usage{
plot_usq(x, y = NULL, main = "", sub = "", xlim = NULL, ylim = NULL,
  ann = par("ann"), ...)
}
\description{
Basic X-Y plotting that follows USQ colour and typography
guidelines.
}
\examples{
plot_usq(1:3)

}
