% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_usq-package.R
\docType{package}
\name{theme.usq}
\alias{theme.usq}
\alias{theme.usq-package}
\title{Provides USQ style typography, colours and graph themes for ggplot2
plots in R.}
\description{
Provides USQ style typography, colours and graph themes for ggplot2
plots in R.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development \url{https://github.com/adamhsparks/theme_usq}}
\item{Report bugs at \url{https://github.com/adamhsparks/theme_usq/issues}}
\item{USQ guidelines (must be logged in) \url{https://www.usq.edu.au/a-brand-hidden/guidelines}}
}
}
\author{
Adam H Sparks
}
