% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_usq.R
\name{theme_usq}
\alias{theme_usq}
\title{A theme based on USQ visual identity and typography guidelines.}
\usage{
theme_usq(base_size = 11, base_family = "Verdana",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A theme based on USQ visual identity and typography guidelines.
}
\examples{
\dontrun{
Discrete values
p1 <- ggplot(mtcars) +
  geom_point(aes(
    x = wt,
    y = mpg,
    colour = factor(gear)
   )) +
  scale_colour_manual(values = usq_palette) +
  facet_wrap(~ am) +
  theme_usq()
p1

# Continous values
df <- reshape2::melt(outer(1:4, 1:4), varnames = c("X1", "X2"))

p2 <- ggplot(df, aes(X1, X2)) + geom_tile(aes(fill = value)) +
  usq_fill_gradient()
p3 <- p2 + geom_point(aes(size = value)) +
  theme_usq()
p3
}

}
