% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R, R/usq_palettes.R
\docType{data}
\name{usq_colours}
\alias{usq_colours}
\alias{usq_palettes}
\alias{usq_colours}
\alias{usq_palettes}
\title{A vector of USQ Colours That Conforms to USQ's Colour Identity}
\format{A \code{vector} of USQ colours from the visual identity
colour palette.}
\usage{
usq_colours

usq_palettes

usq_colours

usq_palettes
}
\description{
A vector of USQ Colours That Conforms to USQ's Colour Identity

A palette of USQ Colours That Conforms to USQ's Colour Identity

A vector of USQ Colours That Conforms to USQ's Colour Identity

A palette of USQ Colours That Conforms to USQ's Colour Identity
}
\keyword{datasets}
