% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R
\name{usq_pal}
\alias{usq_pal}
\title{Return function to interpolate a USQ color palette}
\usage{
usq_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in usq_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a USQ color palette
}
