% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R
\docType{data}
\name{usq_palette}
\alias{usq_palette}
\title{A Palette for Use in Graphing That Conforms to USQ's Colour Identity}
\format{A \code{vector} of USQ colours from the visual identity
colour palette.}
\usage{
usq_palette
}
\description{
A Palette for Use in Graphing That Conforms to USQ's Colour Identity
}
\keyword{datasets}
