% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-usq.R
\name{theme_usq}
\alias{theme_usq}
\title{A theme based on USQ visual identity and typography guidelines.}
\usage{
theme_usq(base_size = 11, base_family = "Verdana",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\description{
A theme based on USQ visual identity and typography guidelines.
}
\examples{

library("ggplot2")
p <- ggplot(mtcars) +
  geom_point(aes(
    x = wt,
    y = mpg,
    colour = factor(gear)
  )) +
  facet_wrap( ~ am)

p + theme_usq()
}
