% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_usq.R
\name{boxplot_usq}
\alias{boxplot_usq}
\title{Basic Boxplots Using USQ Colours and Typography}
\usage{
boxplot_usq(x, main = "", sub = "", xlab = "", ylab = "", col = 1,
  ...)
}
\arguments{
\item{x}{for specifying data from which the boxplots are to be produced.
Either a numeric vector, or a single list containing such vectors. Additional
unnamed arguments specify further data as separate vectors (each
corresponding to a component boxplot). \code{NA}s are allowed in the data.}

\item{main}{Main title. Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label. Optional, if not supplied it will be
blank.}

\item{xlab}{X-axis label. Optional.}

\item{ylab}{Y-axis label. Optional.}

\item{col}{Colour to use as fill and outline for boxes and outlier points as
a digit. Defaults to 1, USQ Charcoal. There are six colours that can be used,
see \strong{Details} for more.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics]{par}}).}
}
\description{
Basic boxplots that follow USQ colour and typography
guidelines that uses (hopefully) sensible defaults.
}
\details{
Colours for use in plots from the USQ Visual Identity Palette as a valid
value for \code{col} include:
\itemize{
\item{1} {- USQ Charcoal}
\item{2} {- USQ Yellow}
\item{3} {- Dark Warm Gray}
\item{4} {- Support Green}
\item{5} {- Secondary Orange}
\item{6} {- Support Blue}
}
}
\examples{
rb <- boxplot_usq(decrease ~ treatment, data = OrchardSprays)
title("Comparing boxplot()s and non-robust mean +/- SD")

}
\seealso{
\code{\link[graphics]{boxplot}}  for full documentation of the basic
boxplot
capabilities
}
