% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_usq.R
\name{hist_usq}
\alias{hist_usq}
\title{Basic Histograms Using USQ Colours and Typography}
\usage{
hist_usq(x, main = "", sub = "", xlab = "", ylab = "Count",
  breaks = "pretty", col = 1, ...)
}
\arguments{
\item{main}{Main title. Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label. Optional, if not supplied it will be
blank.}

\item{xlab}{X-axis label. Optional, if not supplied will be taken from data.}

\item{ylab}{Y-axis label. Optional, if not supplied will default to "Count".}

\item{breaks}{Either \code{pretty}, default, or \code{exact}. Pretty uses
Scott's Rule, whereas exact uses a bin-width of 1 so each value is
represented individually with a bar.}

\item{col}{Colour to use as fill for the bars as a digit. Defaults to 1, USQ
Charcoal. There are six colours that can be used, see \strong{Details} for
more.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics]{par}}).}
}
\description{
Basic histograms that follow USQ colour and typography
guidelines that uses (hopefully) sensible defaults.
}
\details{
Colours for use in plots from the USQ Visual Identity Palette as a valid
value for \code{col} include:
 \itemize{
   \item{1} {- USQ Charcoal}
   \item{2} {- USQ Yellow}
   \item{3} {- Dark Warm Gray}
   \item{4} {- Support Green}
   \item{5} {- Secondary Orange}
   \item{6} {- Support Blue}
 }

 @seealso \code{\link[graphics]{hist}} for full documentation of the basic
histogram capabilities
}
\examples{
library(datasets)

hist_usq(airquality$Ozone,
         col = 3,
         main = "Air Quality",
         xlab = "Ozone",
         breaks = "pretty")

}
\references{
Scott, D.W. (1979) On optimal and data-based histograms. \emph{Biometrika},
66, 605–610.

Hyndman, R. J., The Problem with Sturges’ Rule for Constructing Histograms,
1995, Monash University,
\url{https://www.robjhyndman.com/papers/sturges.pdf}.
}
