% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_usq.R
\name{plot_usq}
\alias{plot_usq}
\title{Basic X-Y Plotting Using USQ Colours and Typography}
\usage{
plot_usq(x, y = NULL, main = "", sub = "", xlab = "", ylab = "",
  xlim = NULL, ylim = NULL, col = 1, ...)
}
\arguments{
\item{x}{Values to be plotted on x-axis.}

\item{y}{Values to be used on y-axis. Optional.}

\item{main}{Main title text}

\item{sub}{Subtitle text below x-axis label}

\item{xlab}{X-axis label. Optional, if blank will default to data supplied
label.}

\item{ylab}{Y-axis label. Optional, if blank will default to data supplied
label if one is available.}

\item{xlim}{Numeric value for x-axis limits. Optional.}

\item{ylim}{Numeric value for y-axis limits. Optional.}

\item{col}{Colour to use as fill for the bars as a digit. Defaults to 1, USQ
Charcoal. There are six colours that can be used, see \strong{Details} for
more.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics]{par}}). The most commonly used argument would be
\code{type} what type of plot should be drawn. Possible types are:
 \itemize{
 \item{"p"} {for \strong{p}oints},
 \item{"l"} for \strong{l}ines,
 \item{"b"} for \strong{b}oth,
 \item{"c"} for the lines part alone of "\strong{b}",
 \item{"o"} for both ‘\strong{o}verplotted’,
 \item{"h"} for ‘\strong{h}istogram’ like (or ‘\strong{h}igh-density’)
 vertical lines,
 \item{"s"} for stair \strong{s}teps,
 \item{"n"} for \strong{n}o plotting.
 }}
}
\description{
Basic X-Y plotting that follows USQ colour and typography
guidelines that uses (hopefully) sensible defaults.
}
\details{
Colours for use in plots from the USQ Visual Identity Palette as a valid
value for \code{col} include:
 \itemize{
   \item{1} {- USQ Charcoal}
   \item{2} {- USQ Yellow}
   \item{3} {- Dark Warm Gray}
   \item{4} {- Support Green}
   \item{5} {- Secondary Orange}
   \item{6} {- Support Blue}
 }
}
\examples{
library(datasets)

plot_usq(airquality$Ozone,
         col = 3,
         main = "Air Quality",
         xlab = "Ozone")

}
\seealso{
\code{\link[graphics]{plot}} for full documentation of the basic plotting
capabilities.
}
