% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_ggplot2_scales.R
\name{scale_fill_usq}
\alias{scale_fill_usq}
\title{ggplot2 fill scale constructor for USQ colours}
\usage{
scale_fill_usq(palette = "primary", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in usq_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{discrete_scale()} or
\code{scale_fill_gradientn()}, used respectively when discrete is TRUE
or FALSE}
}
\description{
ggplot2 fill scale constructor for USQ colours
}
