% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_usq-package.R
\docType{package}
\name{theme.usq-package}
\alias{theme.usq-package}
\title{Provides USQ Style Typography, Colours and Themes for Plots from R}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development \url{https://github.com/adamhsparks/theme.usq}}
\item{Report bugs at \url{https://github.com/adamhsparks/theme.usq/issues}}
\item{USQ guidelines (must be logged in)
\url{https://www.usq.edu.au/a-brand-hidden/guidelines}}
}
}
\author{
Adam H Sparks
}
