% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.usq-package.R, R/theme_usq-package.R
\docType{package}
\name{theme.usq-package}
\alias{theme.usq}
\alias{theme.usq-package}
\title{theme.usq: USQ Themed Plots for Base Graphics and 'ggplot2'}
\description{
Provides University of Southern Queensland (USQ) style,
    typography and colours for plots generated using R base graphics and
    'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamhsparks/theme.usq}
  \item Report bugs at \url{https://github.com/adamhsparks/theme.usq/issues}
}


\strong{Useful links:}
\itemize{
\item{Development \url{https://github.com/adamhsparks/theme.usq}}
\item{Report bugs at \url{https://github.com/adamhsparks/theme.usq/issues}}
\item{USQ guidelines (must be logged in)
\url{https://www.usq.edu.au/a-brand-hidden/guidelines}}
}
}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Other contributors:
\itemize{
  \item Simon Jackson [contributor]
  \item Alicia Allan [contributor]
  \item Di Cook [contributor]
  \item Ross Gayler [contributor]
  \item Holly Kirk [contributor]
  \item Roger Peng [contributor]
  \item Elle Saber [contributor]
}


Adam H Sparks
}
\keyword{internal}
