% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R
\docType{data}
\name{usq_colours}
\alias{usq_colours}
\title{USQ colours}
\format{A \code{vector} of USQ colours from the visual identity
colour palette.}
\usage{
usq_colours
}
\description{
USQ colours
}
\keyword{datasets}
