% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R, R/usq_palettes.R
\name{usq_cols}
\alias{usq_cols}
\alias{usq_cols}
\title{Function to extract USQ colors as hex codes}
\usage{
usq_cols(...)

usq_cols(...)
}
\arguments{
\item{...}{Character names of theme.usq colours}

\item{...}{Character names of theme.usq colours}
}
\description{
Function to extract USQ colors as hex codes

Function to extract USQ colors as hex codes
}
