% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R
\name{usq_cols}
\alias{usq_cols}
\title{Function to extract USQ colours as hex codes}
\usage{
usq_cols(...)
}
\arguments{
\item{...}{Character names of \pkg{theme.usq} colours}
}
\description{
Function to extract USQ colours as hex codes
}
\examples{
# get the hex code for usq yellow
usq_cols("usq yellow")
}
