% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R
\name{usq_pal}
\alias{usq_pal}
\title{Return function to interpolate a USQ colour palette}
\usage{
usq_pal(p = "primary", r = FALSE, a = 1, ...)
}
\arguments{
\item{p}{Character name of palette in \code{usq_palettes()}}

\item{r}{Boolean indicating whether the palette should be reversed}

\item{a}{Alpha transparency value between 1 and 0, where 1 = no transparency
and 0 = transparent}

\item{...}{Additional arguments to pass to \code{colorRampPalette()}}
}
\description{
Return function to interpolate a USQ colour palette
}
