% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_colours.R
\docType{data}
\name{usq_palette}
\alias{usq_palette}
\alias{usq_colour_gradient}
\alias{usq_palette}
\alias{usq_fill_gradient}
\alias{usq_palette}
\alias{usq_fill_gradient}
\alias{usq_palette}
\title{Palettes and functions for the theme.usq package}
\format{A \code{ggplot2} gradient colour function.}
\usage{
usq_colour_gradient()

usq_fill_gradient()

usq_fill_gradient()

usq_palette
}
\description{
Provides colours for graphing using USQ's visual identity
colour palette.
}
\details{
See \code{\link{theme_usq}} for examples.
}
\keyword{datasets}
