% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usq_palettes.R
\docType{data}
\name{usq_palettes}
\alias{usq_palettes}
\title{Palettes of USQ Colours}
\format{\code{Palettes} of USQ colours from the visual identity
colour palette for use in plotting.}
\usage{
usq_palettes
}
\description{
Palettes of USQ Colours
}
\examples{
usq_palettes
}
\keyword{datasets}
